pyproject.toml
[tool.poetry]
name = "doc-room"
version = "0.1.0"
description = ""
authors = ["Manikanta K <manikanta9k7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
